Workaround vscode task config python launch path issue
The latest vscode python extension 2023.16.0 failed to
launch the python process because it expects
the command to be a relative path to the working
directory. However ${command:python.interpreterPath}
is always resolved to an absolute path.
Use a hard coded relative command path for now.

Switch to use Process type to avoid unnecessary overhead and potential complication of launching a shell first.